Remove old Functions project, add CheckZoneEvent
