+++ ApiServer/SilvermineNordic.Functions.Azure/FunctionDefinitions/SensorReadingEvent.cs
