+++ ApiServer/SilvermineNordic.Functions.Azure/FunctionDefinitions/WeatherForecast.cs
