--- ApiServer/SilvermineNordic.Functions/MyInfo.cs
@@ -1,10 +1,0 @@
-﻿
-namespace SilvermineNordic.Functions
-{
-    public class MyInfo
-    {
-        public MyScheduleStatus ScheduleStatus { get; set; }
-
-        public bool IsPastDue { get; set; }
-    }
-}
