--- ApiServer/SilvermineNordic.Functions/MyScheduleStatus.cs
@@ -1,12 +1,0 @@
-﻿
-namespace SilvermineNordic.Functions
-{
-    public class MyScheduleStatus
-    {
-        public DateTime Last { get; set; }
-
-        public DateTime Next { get; set; }
-
-        public DateTime LastUpdated { get; set; }
-    }
-}
