--- ApiServer/SilvermineNordic.Functions/Program.cs
@@ -1,7 +1,0 @@
-using Microsoft.Extensions.Hosting;
-
-var host = new HostBuilder()
-    .ConfigureFunctionsWorkerDefaults()
-    .Build();
-
-host.Run();
