--- ApiServer/SilvermineNordic.Functions/Properties/launchSettings.json
@@ -1,9 +1,0 @@
-{
-  "profiles": {
-    "SilvermineNordic.Functions": {
-      "commandName": "Project",
-      "commandLineArgs": "--port 7047",
-      "launchBrowser": false
-    }
-  }
-}
