--- ApiServer/SilvermineNordic.Functions/Properties/serviceDependencies.json
@@ -1,12 +1,0 @@
-{
-  "dependencies": {
-    "appInsights1": {
-      "type": "appInsights",
-      "connectionId": "APPLICATIONINSIGHTS_CONNECTION_STRING"
-    },
-    "storage1": {
-      "type": "storage",
-      "connectionId": "AzureWebJobsStorage"
-    }
-  }
-}
